+++ app/src/androidTest/java/fi/finhua/android/badminton/ApplicationTest.java
@@ -1,4 +1,4 @@
-package fi.finhua.android.badminton.scoreboard;
+package fi.finhua.android.badminton;
 
 
 import android.app.Application;
