+++ app/src/main/java/fi/finhua/android/badminton/FABBehavior.java
@@ -1,4 +1,4 @@
-package fi.finhua.android.badminton.scoreboard;
+package fi.finhua.android.badminton;
 
 import android.content.Context;
 import android.support.design.widget.AppBarLayout;
