+++ app/src/main/java/fi/finhua/android/badminton/Game.java
@@ -1,4 +1,4 @@
-package fi.finhua.android.badminton.scoreboard;
+package fi.finhua.android.badminton;
 
 import java.util.List;
