+++ app/src/main/java/fi/finhua/android/badminton/GameAdapter.java
@@ -1,4 +1,4 @@
-package fi.finhua.android.badminton.scoreboard;
+package fi.finhua.android.badminton;
 
 import android.support.v7.widget.CardView;
 import android.support.v7.widget.RecyclerView;
