+++ app/src/main/java/fi/finhua/android/badminton/Statistics.java
@@ -1,4 +1,4 @@
-package fi.finhua.android.badminton.scoreboard;
+package fi.finhua.android.badminton;
 
 import android.app.Activity;
 import android.content.Intent;
