+++ app/src/main/java/fi/finhua/android/badminton/Utils.java
@@ -1,4 +1,4 @@
-package fi.finhua.android.badminton.scoreboard;
+package fi.finhua.android.badminton;
 
 import android.content.Context;
 import android.content.res.TypedArray;
